Move span tests to netcoreapp file
