+++ src/System.IO/tests/BinaryReader/BinaryReaderTests.cs
@@ -11,7 +11,7 @@ using Xunit;
 
 namespace System.IO.Tests
 {
-    public class BinaryReaderTests
+    public partial class BinaryReaderTests
     {
         protected virtual Stream CreateStream()
         {
@@ -75,8 +75,6 @@ namespace System.IO.Tests
             Assert.Throws<ObjectDisposedException>(() => binaryReader.Read());
             Assert.Throws<ObjectDisposedException>(() => binaryReader.Read(byteBuffer, 0, 1));
             Assert.Throws<ObjectDisposedException>(() => binaryReader.Read(charBuffer, 0, 1));
-            Assert.Throws<ObjectDisposedException>(() => binaryReader.Read(new Span<byte>(byteBuffer)));
-            Assert.Throws<ObjectDisposedException>(() => binaryReader.Read(new Span<char>(charBuffer)));
             Assert.Throws<ObjectDisposedException>(() => binaryReader.ReadBoolean());
             Assert.Throws<ObjectDisposedException>(() => binaryReader.ReadByte());
             Assert.Throws<ObjectDisposedException>(() => binaryReader.ReadBytes(1));
@@ -130,72 +128,6 @@ namespace System.IO.Tests
                 using (var reader = new BinaryReader(str, new NegEncoding()))
                 {
                     AssertExtensions.Throws<ArgumentOutOfRangeException>("charsRemaining", () => reader.Read(new char[10], 0, 10));
-                }
-            }
-        }
-
-        [Theory]
-        [InlineData(100, 100, 100)]
-        [InlineData(100, 50, 50)]
-        [InlineData(50, 100, 50)]
-        [InlineData(10, 0, 0)]
-        [InlineData(0, 10, 0)]
-        public void Read_ByteSpan(int sourceSize, int destinationSize, int expectedReadLength)
-        {
-            using (var stream = CreateStream())
-            {
-                var source = new byte[sourceSize];
-                new Random(345).NextBytes(source);
-                stream.Write(source, 0, source.Length);
-                stream.Position = 0;
-
-                using (var reader = new BinaryReader(stream))
-                {
-                    var destination = new byte[destinationSize];
-
-                    int readCount = reader.Read(new Span<byte>(destination));
-
-                    Assert.Equal(expectedReadLength, readCount);
-                    Assert.Equal(source.Take(expectedReadLength), destination.Take(expectedReadLength));
-
-                    // Make sure we didn't write past the end
-                    Assert.True(destination.Skip(expectedReadLength).All(b => b == default(byte)));
-                }
-            }
-        }
-
-        [Theory]
-        [InlineData(100, 100, 100)]
-        [InlineData(100, 50, 50)]
-        [InlineData(50, 100, 50)]
-        [InlineData(10, 0, 0)]
-        [InlineData(0, 10, 0)]
-        public void Read_CharSpan(int sourceSize, int destinationSize, int expectedReadLength)
-        {
-            using (var stream = CreateStream())
-            {
-                var source = new char[sourceSize];
-                var random = new Random(345);
-
-                for (int i = 0; i < sourceSize; i++)
-                {
-                    source[i] = (char)random.Next(0, 127);
-                }
-
-                stream.Write(Encoding.ASCII.GetBytes(source), 0, source.Length);
-                stream.Position = 0;
-
-                using (var reader = new BinaryReader(stream, Encoding.ASCII))
-                {
-                    var destination = new char[destinationSize];
-
-                    int readCount = reader.Read(new Span<char>(destination));
-
-                    Assert.Equal(expectedReadLength, readCount);
-                    Assert.Equal(source.Take(expectedReadLength), destination.Take(expectedReadLength));
-
-                    // Make sure we didn't write past the end
-                    Assert.True(destination.Skip(expectedReadLength).All(b => b == default(char)));
                 }
             }
         }
